backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
claude-agent-sdk>=0.1.0
pydantic==2.11.7
python-dotenv==1.1.1
PyPDF2==3.0.1
loguru==0.7.2
python-multipart>=0.0.6
google-genai==1.0.0
